# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "modules Platform API",
    "version": "1.0.0",
    "description": "REST API for interacting the modules platform.",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/your-org/modules-repo"
    },
    "license": {
      "name": "Unlicense",
      "url": "https://unlicense.org/"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}/api",
      "description": "Configurable server",
      "variables": {
        "protocol": {
          "enum": ["http", "https"],
          "default": "http"
        },
        "host": {
          "default": "localhost:3000",
          "description": "Server host and port"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session-based authentication using JWT tokens from Better Auth"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication for programmatic access. Supports 'read' and 'write' scopes."
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/admin/modules": {
      "get": {
        "operationId": "get-admin-modules",
        "summary": "List all modules (admin)",
        "description": "Get a list of all modules with administrative details. Requires admin role.",
        "tags": ["s Admin, Modules"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of modules",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-admin-modules",
        "summary": "Create a new module (admin)",
        "description": "Create a new module directly without going through submission process. Requires admin role.",
        "tags": ["s Admin, Modules"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Module created successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/modules": {
      "get": {
        "operationId": "get-modules",
        "summary": "List modules",
        "description": "Get a list of published modules with optional filtering, search, and sorting. Supports both session-based and API key authentication.",
        "tags": ["s Modules"],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "description": "Filter by category"
            },
            "required": false,
            "description": "Filter by category"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string",
              "description": "Search query for module names and descriptions"
            },
            "required": false,
            "description": "Search query for module names and descriptions"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string",
              "enum": ["featured", "recommended", "recent"],
              "description": "Pre-defined filter type"
            },
            "required": false,
            "description": "Pre-defined filter type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of modules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "modules": {
                      "type": "array",
                      "items": {},
                      "description": "Array of module objects"
                    }
                  },
                  "required": ["modules"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/modules/{id}": {
      "get": {
        "operationId": "get-modules-{id}",
        "summary": "Get module details",
        "description": "Retrieve information about a specific module including metadata, compatibility, features, and optionally releases. Supports both session-based and API key authentication.",
        "tags": ["s Modules"],
        "parameters": [
          {
            "in": "query",
            "name": "includeReleases",
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "description": "Include release information"
            },
            "required": false,
            "description": "Include release information"
          },
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Module ID"
            },
            "required": true,
            "description": "Module ID",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved module details with full metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "module": {
                      "description": "Module object with full details"
                    }
                  },
                  "required": ["module"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "operationId": "get-search",
        "summary": "Advanced module search",
        "description": "Search modules with advanced filtering, sorting, and pagination. Supports full-text search across module names, descriptions, and authors with filtering options.",
        "tags": ["s Search"],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string",
              "description": "Search query"
            },
            "required": false,
            "description": "Search query"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "description": "Filter by category"
            },
            "required": false,
            "description": "Filter by category"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": ["relevance", "downloads", "rating", "updated", "name"],
              "description": "Sort order"
            },
            "required": false,
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "description": "Sort direction"
            },
            "required": false,
            "description": "Sort direction"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "description": "Page number for pagination"
            },
            "required": false,
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "description": "Items per page"
            },
            "required": false,
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "minRating",
            "schema": {
              "type": "number",
              "description": "Minimum rating filter"
            },
            "required": false,
            "description": "Minimum rating filter"
          },
          {
            "in": "query",
            "name": "isOpenSource",
            "schema": {
              "type": "boolean",
              "description": "Filter by open source status"
            },
            "required": false,
            "description": "Filter by open source status"
          },
          {
            "in": "query",
            "name": "rootMethod",
            "schema": {
              "type": "string",
              "description": "Filter by root method"
            },
            "required": false,
            "description": "Filter by root method"
          },
          {
            "in": "query",
            "name": "androidVersion",
            "schema": {
              "type": "string",
              "description": "Filter by Android version compatibility"
            },
            "required": false,
            "description": "Filter by Android version compatibility"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved search results with metadata and pagination info",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-search",
        "summary": "Search suggestions",
        "description": "Get intelligent search suggestions and autocomplete for module names, authors, and categories. Helps users discover relevant content as they type.",
        "tags": ["s Search"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved search suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": ["module", "author", "category"]
                          },
                          "value": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "count": {
                            "type": "number",
                            "nullable": true
                          }
                        },
                        "required": ["type", "value", "label"]
                      },
                      "description": "Search suggestions"
                    }
                  },
                  "required": ["suggestions"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}